fix(api): repair pause/resume, wrong error codes and spurious 500s
Three defects found by following the documentation as a new developer would.

Pausing or resuming a monitor always failed with a 500, from the UI as well as
the API: inside a CASE expression Postgres cannot infer a bare parameter's type
and defaults it to text, which the smallint status column rejects. The feature
had never worked.

GET /auth/me, POST /auth/change-password and both /workspaces endpoints
accepted an API key at the guard and then dereferenced the user it does not
have, producing a 500 and a stack trace per request. They return 403 now, which
is what the documentation already claimed.

Errors raised by Fastify and its plugins carry a status but no domain code, and
everything below 500 was labelled validation_error — so a rate-limited caller
got 429 with code "validation_error" while the documented "rate_limited" was
never emitted. The code is now derived from the status.

Also throttles the API-key last-used write to at most once per minute; it ran
on every authenticated request, rewriting the same row under load.

+++ packages/db/src/repositories/monitor-repository.ts
@@ -314,8 +314,11 @@ export class MonitorRepository {
   /** Pause/resume a monitor and mirror the flag onto its assignments. */
   async setEnabled(monitorId: string, enabled: boolean): Promise<void> {
     await this.db.query(
+      // The status casts are required: inside CASE, Postgres cannot infer the
+      // type of a bare parameter and defaults it to text, which fails against
+      // the smallint column.
       `UPDATE monitors
-       SET enabled = $2, status = CASE WHEN $2 THEN $3 ELSE $4 END
+       SET enabled = $2, status = CASE WHEN $2 THEN $3::smallint ELSE $4::smallint END
        WHERE id = $1`,
       [
         monitorId,
